% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cin_metrics.R
\name{cna}
\alias{cna}
\title{Copy Number Aberration
Calculates the number of copy number aberrations
Nearly identical to countingBreakPoints, except this one calculates breaks as adjacent segments that have a difference in segment means of >= 0.2.}
\usage{
cna(
  cnvData,
  segmentMean = (log(1.7, 2) - 1),
  numProbes = NA,
  segmentDistance = 0.2,
  minSegSize = 10
)
}
\arguments{
\item{cnvData}{dataframe containing following columns: Sample, Start, End, Num_Probes, Segment_Mean}

\item{segmentMean}{numerical value for the minimum segment_mean cutoff/ threshold. Default is 0.2}

\item{numProbes}{Number of Probes}

\item{segmentDistance}{Segment distance threshold}

\item{minSegSize}{Minimum segment size}
}
\value{
Number of copy number aberrations between segments
}
\description{
Copy Number Aberration
Calculates the number of copy number aberrations
Nearly identical to countingBreakPoints, except this one calculates breaks as adjacent segments that have a difference in segment means of >= 0.2.
}
