% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cin_metrics.R
\name{fga}
\alias{fga}
\title{Fraction Genome Altered
Fraction Genome Altered looks at the fraction of the genome that deviates from a diploid state}
\usage{
fga(cnvData, segmentMean = 0.2, numProbes = NA, genomeSize = 2873203431)
}
\arguments{
\item{cnvData}{dataframe containing following columns: Sample, Start, End, Num_Probes, Segment_Mean}

\item{segmentMean}{numerical value for the minimum segment_mean cutoff/ threshold. Default is 0.2}

\item{numProbes}{Number of Probes}

\item{genomeSize}{Size of the genome derived from Affymetrix 6.0 array probe. Default is 2873203431 calculated based on hg38 **excluding sex chromosomes**}
}
\value{
Fraction of the genome altered
}
\description{
Fraction Genome Altered
Fraction Genome Altered looks at the fraction of the genome that deviates from a diploid state
}
