% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maskCNV_BRCA}
\alias{maskCNV_BRCA}
\title{Breast Cancer Data from TCGA
Data Release 25.0
GDC Product: Data
Release Date: July 22, 2020
Masked Copy Number variation data for Breast Cancer for 10 unique samples selected randomly from TCGA}
\format{
An object of class dataframe
}
\source{
\url{https://portal.gdc.cancer.gov/}
}
\usage{
data(maskCNV_BRCA)
}
\description{
Breast Cancer Data from TCGA
Data Release 25.0
GDC Product: Data
Release Date: July 22, 2020
Masked Copy Number variation data for Breast Cancer for 10 unique samples selected randomly from TCGA
}
\examples{
data(maskCNV_BRCA)
\donttest{tai <- tai(maskCNV_BRCA)}
}
\references{
Koboldt, D., Fulton, R., McLellan, M. et al. (2012) Nature 490, 61–70
\url{https://www.nature.com/articles/nature11412}
}
\keyword{dataset}
